Actually run some tests!
This is not a very good solution. We need a way to run all tests in one
command, the exit code of which will determine the status of the build - having
a long string of commands chained with '&&' is a poor way of implementing this.

This can likely be achieved by using CTest and 'make test'